mmaprototype: move unused helper functions to constraint_unused.go
Previously, constraint.go included a few helper functions intended for future
production use but not currently in use. This commit moves those helpers into
their own dedicated file to make constraint.go smaller.
