+++ pkg/kv/kvserver/allocator/mmaprototype/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "cluster_state_rebalance_stores.go",
         "constraint.go",
         "constraint_matcher.go",
+        "constraint_unused.go",
         "doc.go",
         "load.go",
         "memo_helper.go",
